X4←X6

+++ Soulver/SVRSolver.m
@@ -83,7 +83,7 @@ NSCharacterSet *SVRSolverTextAttachmentCharacterSet = nil;
   [self __step5_styleAndTag:input styles:textStyles];
   outputLength = [[input string] length];
   if (inputLength != outputLength) {
-    XPLogPause2(@"SVRSolver solveAttributedString: String changed length: %ld->%ld", inputLength, outputLength);
+    XPLogPause2(@"SVRSolver solveAttributedString: String changed length: %d->%d", (int)inputLength, (int)outputLength);
   }
   [input autorelease];
 }
